update: update latest shadcn, and fix error corepack
