+++ package.json
@@ -36,7 +36,7 @@
     "clsx": "^2.1.1",
     "embla-carousel-autoplay": "^8.2.0",
     "embla-carousel-react": "^8.2.0",
-    "lucide-react": "^0.451.0",
+    "lucide-react": "^0.511.0",
     "motion": "^12.7.4",
     "next": "15.3.2",
     "next-intl": "^3.17.6",
@@ -45,7 +45,7 @@
     "react": "19.1.0",
     "react-dom": "19.1.0",
     "react-hook-form": "^7.56.2",
-    "tailwind-merge": "^3.2.0",
+    "tailwind-merge": "^3.3.0",
     "tailwindcss-animate": "^1.0.7",
     "zod": "^3.24.4"
   },
@@ -72,6 +72,7 @@
     "prettier": "3.3.3",
     "tailwindcss": "^4.1.4",
     "ts-node": "^10.9.2",
+    "tw-animate-css": "^1.3.0",
     "typescript": "^5",
     "vitest": "^2.1.3"
   },
